+++ app/jobs/access_request_satisfaction_survey_job.rb
@@ -1,0 +1,7 @@
+class AccessRequestSatisfactionSurveyJob < ApplicationJob
+  queue_as :default
+
+  def perform(*args)
+    JwtApiEntreprise::Operation::AccessRequestSatisfactionSurvey.call
+  end
+end
