--- app/jobs/jwt_satisfaction_survey_job.rb
@@ -1,7 +1,0 @@
-class JwtSatisfactionSurveyJob < ApplicationJob
-  queue_as :default
-
-  def perform(*args)
-    JwtSatisfactionSurvey::Operation::Create.call
-  end
-end
