--- spec/jobs/jwt_satisfaction_survey_job_spec.rb
@@ -1,13 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe JwtSatisfactionSurveyJob do
-  before do
-    expect(JwtSatisfactionSurvey::Operation::Create).to receive(:call)
-  end
-
-  subject { described_class }
-
-  it 'calls the operation' do
-    subject.perform_now
-  end
-end
